% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tens.tpls.R
\name{tpls}
\alias{tpls}
\title{Tensor PLS-like regression}
\usage{
tpls(
  x,
  y,
  ncomp = NULL,
  m = NULL,
  minv = NULL,
  mode = "regression",
  center = TRUE,
  matrix_output = TRUE,
  bpparam = NULL
)
}
\arguments{
\item{x}{Tensor input.}

\item{y}{Tensor input.}

\item{ncomp}{The estimated number of components. ncomp must be explicitly set
as an integer in tpls.}

\item{m}{A function which applies an orthogonal tensor tubal transform.}

\item{minv}{The inverse of m.}

\item{mode}{Currently supports tensor analogues of canonical ("canonical"),
regression ("regression"), and svd ("tsvdm") PLS variants. Defaults to
"regression".}

\item{center}{If set to false, the data tensor will not be centralized into
Mean Deviation Form (see Mor et al. 2022). By default, the mean horizontal
slice of the input tensor(s) are subtracted, so that all of the horizontal
slices sum to 0, analgous to centering matrix data.}

\item{matrix_output}{Note: ONLY AFFECTS THE OUTPUT IN "tsvdm" MODE.
Collects the top singular vectors across the tensor, organized by the
magnitude of the corresponding singular vector, and place them into the
columns of a matrix. Corresponds to the 'matrix compression' type of scheme
described in Mor et al. 2022.}

\item{bpparam}{A \link[BiocParallel]{BiocParallelParam-class} object
indicating the type of parallelisation. Does not have any effect if transform
functions explicitly set using \code{m}, \code{minv}.}
}
\value{
A list containing:
\describe{
\item{ncomp}{The number of components used.}
\item{x}{The centered input tensor X.}
\item{y}{The centered input tensor Y.}
\item{mode}{The PLS mode used: "canonical", "regression", or "tsvdm".}
\item{x_loadings}{Matrix of X loadings (p x ncomp), or loadings tensor
if \code{mode = "tsvdm"} AND \code{matrix_output = FALSE}.}
\item{y_loadings}{Matrix of Y loadings (q x ncomp), or loadings tensor
if \code{mode = "tsvdm"} AND \code{matrix_output = FALSE}.}
\item{x_projected}{Matrix of X projections (n x ncomp), or proections
tensor when \code{mode = "tsvdm"} AND \code{matrix_output = FALSE}.}
\item{y_projected}{Matrix of Y projections (n x ncomp), or proections
tensor when \code{mode = "tsvdm"} AND \code{matrix_output = FALSE}.}
\item{features}{Vector of selected feature indices per component (only
when \code{matrix_output = TRUE}).}
\item{faces}{Vector of selected tensor face indices per component (only
when \code{matrix_output = TRUE}).}
}
}
\description{
Developed @ Melbourne Integrative Genomics. Intuition: Fourier-based m
transforms compact the data across time points, but preserve the information
association for each sample and each feature. In doing so, pls works by
calculating the singular vectors associated with a 'global' (across time
points) top singular value, and deflating just that relevant face.
}
\author{
Brendan Lu
}
